start porting Prep import over to an abstracted Prep Manager model
