+++ app/controllers/admin/preparations_controller.rb
@@ -1,9 +1,9 @@
 class Admin::PreparationsController < AdminController
-  def import
-    Preparation.import(params[:file])
-    flash[:success] = "Preparations Updated"
-    redirect_to admin_preparations_path
-  end
+  # def import
+  #   Preparation.import(params[:file])
+  #   flash[:success] = "Preparations Updated"
+  #   redirect_to admin_preparations_path
+  # end
 
   def preparations
     @preparations = Preparation.all
